+++ Assets/NeoForge/Scripts/Dialogue.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 9a5d16d5a8665594aa32726655d029cd
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
