+++ Assets/NeoForge/Scripts/Dialogue/Editor.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: ecb78ae7825c9b14f8cb50e019276262
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
