+++ Assets/NeoForge/Scripts/SaveSystem.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 8304d7e9439355649a2d28b9b8e3c681
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
