+++ Assets/_InDevelopment/Chase/RefactoredScripts/AudioDatabase.cs
@@ -1,7 +1,7 @@
 ﻿using AYellowpaper.SerializedCollections;
 using UnityEngine;
 
-namespace NeoForge.Dialogue
+namespace NeoForge.Dialogue.Audio
 {
     [CreateAssetMenu(fileName = "AudioDatabase", menuName = "AudioDatabase")]
     public class AudioDatabase : ScriptableObject
