+++ Assets/_InDevelopment/Chase/RefactoredScripts/ConversantType.cs
@@ -1,0 +1,9 @@
+﻿namespace NeoForge.Dialogue.Helper
+{
+    public enum ConversantType
+    {
+        Player,
+        Conversant,
+        Voice,
+    }
+}
