+++ Assets/_InDevelopment/Chase/RefactoredScripts/ConversantType.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: ef9fe054b9964d41a81369bea5529bfe
+timeCreated: 1718025783
