+++ Assets/_InDevelopment/Chase/RefactoredScripts/DialogueChain.cs
@@ -1,0 +1,11 @@
+﻿using System;
+using System.Collections.Generic;
+
+namespace NeoForge.Dialogue.Helper
+{
+    [Serializable]
+    public class DialogueChain
+    {
+        public List<DialogueData> dialogues = new();
+    }
+}
