+++ Assets/_InDevelopment/Chase/RefactoredScripts/DialogueData.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: fefc2c5acf99460d8b7f34ff15ddc38c
+timeCreated: 1718025783
