+++ Assets/_InDevelopment/Chase/RefactoredScripts/DialogueVariables.cs
@@ -1,0 +1,16 @@
+﻿namespace NeoForge.Dialogue.Helper
+{
+    public static class DialogueVariables
+    {
+        public static readonly string ID_MARKER = "ID: ";
+        public static readonly string CONDITIONAL_MARKER = "Conditions";
+        public static readonly string CHANGES_MARKER = "Changes";
+        public static readonly string DIALOGUE_MARKER = "Dialogue";
+        public static readonly string PLAYER_MARKER = "Player: ";
+        public static readonly string VOICE_MARKER = "Voice: ";
+        public static readonly string LEADS_TO_MARKER = "Leads To:";
+        public static readonly string EVENT_MARKER = "*";
+        public static readonly string VARIATION_MARKER = "Variation";
+        public static readonly string MUSIC_MARKER = "Cue Audio";
+    }
+}
