+++ Assets/_InDevelopment/Chase/RefactoredScripts/StateRequirement.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: bc8d2c8c8b034f41848e270e9e35d475
+timeCreated: 1718025783
