+++ Assets/_InDevelopment/Chase/RefactoredScripts/TriggerDialogueOnStart.cs
