+++ Assets/_InDevelopment/Chase/RefactoredScripts/VisibleBasedOnWorldState.cs
@@ -1,6 +1,6 @@
 ﻿using UnityEngine;
 
-namespace NeoForge.Dialogue
+namespace NeoForge.Dialogue.UI
 {
     public class VisibleBasedOnWorldState : MonoBehaviour
     {
